Added the base functionality of the C program
This program simulates people drawing names from a hat in order to determine who they are exchanging a gift with. The user enters how many people are participating and the first and last name of each person. The program will then match everyone with the person they are going to get a gift for and then print the results.